oprava funkcii a test
Spravne nakonfigurovane funkcia na init usartu ad prevodnika aj
preruseni. Program komunikuje aj cita z ad prevodnika. Program pise to
pismeno co mu bolo poslane.

+++ src/vrs_cv5.h
@@ -13,7 +13,7 @@
 
 //definicie globalnych premennych
 uint16_t value;
-uint8_t rec_data;
+uint16_t rec_data;
 
 // definicie funkcii
 void adc_init();
